28/8

+++ 1_Authentication/server/app/api/v1/users.py
@@ -1,5 +1,5 @@
 from fastapi import APIRouter, Depends, HTTPException
-
+from pydantic import BaseModel
 from api.deps import SessionDep, RoleChecker, check_permissions
 from schemas.users import UserCreate, User, UserUpdate
 from services.users import UserService
@@ -15,7 +15,7 @@ async def read_users(
     offset: int = 0,
     limit: int = 100,
 ):  
-    # Kiểm tra quyền truy cập
+
     kwargs = await check_permissions(current_user, action="get_multi", owner_field="id")
     
     result = await user_service.get_users_service(session=session, offset=offset, limit=limit, kwargs=kwargs)
@@ -25,13 +25,13 @@ async def read_users(
   
 
 @router.get("/{user_id}/", response_model=User)
-async def read_user(
+async def get_user(
     user_id: int,
     session: SessionDep,
     current_user: User = Depends(RoleChecker(["admin", "user"])), 
 ):
 
-    kwargs = await check_permissions(current_user, action="get", owner_field="id")
+    kwargs = await check_permissions(current_user, action="get", owner_field="id", id=user_id)
     result = await user_service.get_user_service(session=session, user_id=user_id, kwargs=kwargs)
     if isinstance(result, Exception):
         raise HTTPException(status_code=result.code, detail={"message": result.message, "code": result.code})
@@ -46,7 +46,6 @@ async def create_user(
 ):
 
     kwargs = await check_permissions(current_user, action="create", obj_in=user_in.dict(), owner_field="id")
-    
     result = await user_service.create_user_service(session=session, user_in=user_in, kwargs=kwargs)
     if isinstance(result, Exception):
         raise HTTPException(status_code=result.code, detail={"message": result.message, "code": result.code})
@@ -57,10 +56,9 @@ async def update_user(
     user_id: int,
     session: SessionDep,
     user_in: UserUpdate,
-    current_user: User = Depends(RoleChecker(["admin", "user"])) 
+    current_user: User = Depends(RoleChecker(["user, admin"])) 
 ):
- 
-    kwargs = await check_permissions(current_user, action="update", owner_field="id")
+    kwargs = await check_permissions(current_user, action="update", owner_field="id", id=user_id)
     result = await user_service.update_user_service(session=session, user_id=user_id, user_in=user_in, kwargs=kwargs)
     if isinstance(result, Exception):
         raise HTTPException(status_code=result.code, detail={"message": result.message, "code": result.code})
@@ -72,9 +70,9 @@ async def delete_user(
     session: SessionDep,
     current_user: User = Depends(RoleChecker(["admin"])) 
 ):
-
     kwargs = await check_permissions(current_user, action="delete", owner_field="id")
-    
+    if user_id == current_user.id:
+        raise HTTPException(status_code=400, detail="You can't delete yourself")
     result = await user_service.delete_user_service(session=session, user_id=user_id, kwargs=kwargs)
     if isinstance(result, Exception):
         raise HTTPException(status_code=result.code, detail={"message": result.message, "code": result.code})
